Fix and add integration tests to allow subunits
allow subunits in the amount field.

+++ app/services/create_charge.rb
@@ -34,7 +34,7 @@ class CreateCharge
   def amount=(amount)
     @amount =
       begin
-        Integer(amount.strip) * 100
+        Float(amount.strip) * 100
       rescue ArgumentError
         0
       end
